+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `reports.schemas.js` Updated `reportResponseSchema` to include property definition for `comments` array
+
 ## [1.0.0] - 2024-02-13
 
 ### Added